myhelloapp/requirements.txt
Django==1.9.6
django-registration-redux==1.3
django-twitter-bootstrap==3.3.0
Pillow==3.3.1
